requirements.txt
livekit-agent
livekit-plugins-groq
livekit-plugins-silero
livekit-plugins-elevenlabs
python-dotenv